[GenAI]: make model compatible with pdf and txt files

+++ gemini_assistant/deploy.py
@@ -1,3 +1,13 @@
+# created by: mintRaven-05
+# dated: 29 sept 2024
+# modified 30 sept 2024
+#           This module is responsible for deploying
+#           the gemini assistant over the localhost
+#           using FastAPI. I contains route to /chat
+#           /image and /doc for processing normal text
+#           chats, images and documents respectively.
+#
+# -----------------------------------------------------------------------------------
 import os
 import io
 import base64
@@ -8,17 +18,29 @@ from docs import *
 from PIL import Image
 from pydantic import BaseModel
 from fastapi import FastAPI, File, Form, UploadFile, HTTPException, Request, status
+# -----------------------------------------------------------------------------------
 
 app = FastAPI()
+
+ALLOWED_FILE_TYPES = {
+    "application/pdf": ".pdf",
+    "text/plain": ".txt",
+}
 
 
 class ImageReq(BaseModel):
     image: str
 
 
+# -----------------------------------------------------------------------------------
+
+
 @app.get("/")
 def index():
     return {"health": "ok", "status": "working"}
+
+
+# -----------------------------------------------------------------------------------
 
 
 @app.post("/chat")
@@ -32,6 +54,9 @@ async def pchat(request: Request):
         raise HTTPException(status_code=400, detail=str(e))
 
 
+# -----------------------------------------------------------------------------------
+
+
 @app.post("/image")
 def pimage(request: ImageReq, query: str = Form(...)):
     try:
@@ -43,10 +68,25 @@ def pimage(request: ImageReq, query: str = Form(...)):
         raise HTTPException(status_code=400, detail=str(e))
 
 
+# -----------------------------------------------------------------------------------
+
+
 @app.post("/doc")
 async def pdoc(doc: UploadFile = File(...), query: str = Form(...)):
+    temp_file_path = None
+
     try:
-        with tempfile.NamedTemporaryFile(delete=False, suffix=".pdf") as temp_file:
+        print("Received MIME type:", doc.content_type)
+
+        if doc.content_type not in ALLOWED_FILE_TYPES:
+            raise HTTPException(
+                status_code=400,
+                detail="Invalid file type! Only PDF and TXT files are allowed.",
+            )
+
+        with tempfile.NamedTemporaryFile(
+            delete=False, suffix=ALLOWED_FILE_TYPES[doc.content_type]
+        ) as temp_file:
             temp_file.write(await doc.read())
             temp_file_path = temp_file.name
 
@@ -57,5 +97,8 @@ async def pdoc(doc: UploadFile = File(...), query: str = Form(...)):
         raise HTTPException(status_code=500, detail=str(e))
 
     finally:
-        if os.path.exists(temp_file_path):
+        if temp_file_path and os.path.exists(temp_file_path):
             os.remove(temp_file_path)
+
+
+# -----------------------------------------------------------------------------------
